qml: Remove the swipe-to-close feature of the keyboard
There have been numerous problems with this interaction over the years,
and the implementation to make swipe to close work is a bit spread around.
One consistent issue has been the keyboard getting stuck partly closed, and
swiping upward does nothing, while in some cases swiping upward could also
cause the keyboard to be moved partially off the screen. This implementation
would also be in conflict of possible implementation of skeyer swipe support
in the future. So just remove it for now.